+++ CHANGES.md
@@ -3,4 +3,5 @@ CHANGELOG
 
 v4.5-r1 (2025-04-30)
 ------------------
+* [FEATURE] Add admin setting to control if the semester field should be visible and editable in the course settings
 * Moodle 4.5 compatible version