+++ CHANGES.md
@@ -1,6 +1,10 @@
 CHANGELOG
 =========
 
+v4.5-r2 (2025-07-09)
+------------------
+* [FEATURE] Replace admin setting visibleincoursettings by capability edit_semester
+
 v4.5-r1 (2025-04-30)
 ------------------
 * [FEATURE] Add admin setting to control if the semester field should be visible and editable in the course settings